[refactor]auth-userのフォルダ整理
