+++ server/src/auth-user/auth-user.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AuthUserController } from './auth-user.controller';
-import { GoogleStrategy } from './google.strategy';
-import { AuthUserService } from './auth-user.service';
-import { JwtStrategy } from './jwt.strategy';
+import { AuthUserController } from './controllers/auth-user.controller';
+import { GoogleStrategy } from './strategies/google.strategy';
+import { AuthUserService } from './services/auth-user.service';
+import { JwtStrategy } from './strategies/jwt.strategy';
 
 @Module({
-    controllers: [AuthUserController],
-    providers: [GoogleStrategy, JwtStrategy, AuthUserService],
+  controllers: [AuthUserController],
+  providers: [GoogleStrategy, JwtStrategy, AuthUserService],
 })
-export class AuthUserModule { }
+export class AuthUserModule {}
